# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bake Delight Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: visible
          in: query
          schema:
            type: boolean
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
  /products/{id}/toggle-visibility:
    patch:
      operationId: toggleProductVisibility
      tags: [products]
      summary: Toggle product visibility (admin)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}/toggle-availability:
    patch:
      operationId: toggleProductAvailability
      tags: [products]
      summary: Toggle product availability (admin)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products by category
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  # ── Orders ────────────────────────────────────────────────────────────────
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders (admin)
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/OrderStatus"
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order (customer checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /orders/calendar:
    get:
      operationId: getOrdersCalendar
      tags: [orders]
      summary: Orders grouped by delivery date for calendar view
      parameters:
        - name: month
          in: query
          schema:
            type: string
            description: "YYYY-MM"
      responses:
        "200":
          description: Calendar entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEntry"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/check-capacity:
    get:
      operationId: checkDeliveryCapacity
      tags: [orders]
      summary: Check if a delivery date/time slot has capacity
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: timeSlot
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Capacity status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CapacityStatus"

  # ── Banners ───────────────────────────────────────────────────────────────
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, slug, sortOrder, createdAt]
...
    ProductVariant:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          description: "e.g. size, flavor, frosting"
        options:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              priceAdjustment:
                type: number
            required: [label, priceAdjustment]
      required: [name, type, options]
    ProductAddon:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
      required: [name, price]
    Product:
      type: object
      properties:
        id:
          type: integer
        categoryId:
          type: integer
          nullable: true
        category:
          $ref: "#/components/schemas/Category"
          nullable: true
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        basePrice:
          type: number
        imageUrls:
          type: array
          items:
            type: string
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        addons:
          type: array
          items:
            $ref: "#/components/schemas/ProductAddon"
        allowCustomMessage:
          type: boolean
        isVisible:
          type: boolean
        isAvailable:
          type: boolean
        orderCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, slug, basePrice, imageUrls, variants, addons, allowCustomMessage, isVisible, isAvailable, orderCount, createdAt]
    CreateProductBody:
      type: object
      properties:
        categoryId:
          type: integer
          nullable: true
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        basePrice:
          type: number
        imageUrls:
          type: array
          items:
            type: string
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        addons:
          type: array
          items:
            $ref: "#/components/schemas/ProductAddon"
        allowCustomMessage:
          type: boolean
        isVisible:
          type: boolean
        isAvailable:
          type: boolean
      required: [name, slug, basePrice]

    # ── Order ─────────────────────────────────────────────────────────────────
    OrderStatus:
      type: string
      enum: [pending, confirmed, in_baking, out_for_delivery, completed, cancelled]
    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        productImageUrl:
          type: string
          nullable: true
        quantity:
          type: integer
        unitPrice:
          type: number
        selectedVariants:
          type: object
          additionalProperties:
            type: string
        selectedAddons:
          type: array
          items:
            type: string
        customMessage:
          type: string
          nullable: true
        subtotal:
          type: number
      required: [productId, productName, quantity, unitPrice, selectedVariants, selectedAddons, subtotal]
    Order:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
          nullable: true
        deliveryAddress:
          type: string
        deliveryDate:
          type: string
          format: date
        deliveryTimeSlot:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/OrderStatus"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        discountAmount:
          type: number
        totalAmount:
          type: number
        couponCode:
          type: string
          nullable: true
        whatsappMessageSent:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, customerName, customerPhone, deliveryAddress, deliveryDate, status, items, subtotal, discountAmount, totalAmount, whatsappMessageSent, createdAt]
    CreateOrderBody:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
          nullable: true
        deliveryAddress:
          type: string
        deliveryDate:
          type: string
          format: date
        deliveryTimeSlot:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        couponCode:
          type: string
          nullable: true
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: integer
              selectedVariants:
                type: object
                additionalProperties:
                  type: string
              selectedAddons:
                type: array
                items:
                  type: string
              customMessage:
                type: string
                nullable: true
            required: [productId, quantity, selectedVariants, selectedAddons]
      required: [customerName, customerPhone, deliveryAddress, deliveryDate, items]
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"
      required: [status]
...
    CalendarEntry:
      type: object
      properties:
        date:
          type: string
          format: date
        orderCount:
          type: integer
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
      required: [date, orderCount, orders]
    CapacityStatus:
      type: object
      properties:
        available:
          type: boolean
        remainingSlots:
          type: integer
        isBlackout:
          type: boolean
      required: [available, remainingSlots, isBlackout]

    # ── Banner ─────────────────────────────────────────────────────────────────
...